Blob cache dummy moved out of solo to avoid circular dependencies
